requirements.txt
aiogram==3.0.0b8
tenacity==8.2.2
tortoise-orm==0.19.3
asyncmy==0.2.8
aerich==0.7.2
httpx==0.24.1
jdatetime==4.1.1
qrcode==7.4.2
pillow==10.0.0
python-dateutil==2.8.2
apscheduler==3.10.1
redis==4.6.0
python-decouple==3.8
python-dotenv==1.0.0
aiohttp
